Extract out even further for android
