+++ Kolektivo/android/app/src/main/java/org/celo/mobile/MainApplication.java
@@ -42,7 +42,6 @@ public class MainApplication
     protected List<ReactPackage> getPackages() {
       @SuppressWarnings("UnnecessaryLocalVariable")
       List<ReactPackage> packages = new PackageList(this).getPackages();
-      packages.add(new CapsuleSignerPackage());
       return packages;
     }
