+++ capsule/android/build.gradle
@@ -63,6 +63,7 @@ dependencies {
   // For > 0.71, this will be replaced by `com.facebook.react:react-android:$version` by react gradle plugin
   //noinspection GradleDynamicVersion
   implementation "com.facebook.react:react-native:+"
+  implementation files('../signer.aar')
 }
 
 if (isNewArchitectureEnabled()) {
