+++ capsule/android/src/main/java/com/capsule/reactnativewallet/CapsuleSignerPackage.java
@@ -1,4 +1,4 @@
-package org.celo.mobile;
+package com.capsule.reactnativewallet;
 
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
@@ -11,12 +11,16 @@ import java.util.List;
 public class CapsuleSignerPackage implements ReactPackage {
 
   @Override
-  public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
+  public List<ViewManager> createViewManagers(
+    ReactApplicationContext reactContext
+  ) {
     return Collections.emptyList();
   }
 
   @Override
-  public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
+  public List<NativeModule> createNativeModules(
+    ReactApplicationContext reactContext
+  ) {
     List<NativeModule> modules = new ArrayList<>();
 
     modules.add(new CapsuleSignerModule(reactContext));
